requirements.txt
Flask
Pillow
numpy
pandas
colorspacious
